backend/requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
python-dotenv==1.2.1
pytokens==0.3.0
sqlparse==0.5.4
